# icode/server-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "yonder博客的后端API服务",
        "title": "yonder blog api aservice",
        "termsOfService": "https://github.com/swaggo/swag",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8030",
    "basePath": "/",
    "paths": {
...
        "/api/v1/user/signin": {
            "post": {
                "description": "检查用户是否存在，密码是否正确，如果正常，返回用户token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登陆",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/req.SignInReq"
                        }
                    }
                ],
                "responses": {
                    "111001": {
                        "description": "参数错误",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SignInResp"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/signup": {
            "post": {
                "description": "创建新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/req.SignupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/renderx.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "renderx.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "result": {
                    "$ref": "#/definitions/renderx.Result"
                }
            }
        },
        "renderx.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "req.SignInReq": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        },
        "req.SignupReq": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        },
...
        "resp.SignInResp": {
            "type": "object",
            "properties": {
                "user_token": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}